`feature`: metatable support; `refactor`: split code into modules
