+++ src/stdlib/io/std_file_kind.rs
@@ -1,0 +1,9 @@
+use gc_arena::Collect;
+
+#[derive(Collect, Clone, Copy)]
+#[collect(no_drop)]
+pub enum StdFileKind {
+    Stdin,
+    Stdout,
+    Stderr,
+}
