+++ src/stdlib/mod.rs
@@ -6,6 +6,6 @@ mod string;
 mod table;
 
 pub use self::{
-    base::load_base, coroutine::load_coroutine, io::load_io, math::load_math, string::load_string,
+    base::load_base, coroutine::load_coroutine, io::{load_io, IoFile}, math::load_math, string::load_string,
     table::load_table,
 };
