--- tests/scripts/io-test-output
@@ -1,1 +1,0 @@
-Hello, world!
