+++ tests/scripts/io.lua
@@ -1,7 +1,2 @@
 io.input("./tests/scripts/io-test-input")
 io.output("./tests/scripts/io-test-output")
-
-do
-    io.write("Hello, world!")
-    print(io.read("l"))
-end
